QuantumMechanics

Documentation for QuantumMechanics.

QuantumMechanics.CompositeBasis

QuantumMechanics.FockBasis

QuantumMechanics.GenericBasis

QuantumMechanics.SpinBasis

Base.length

QuantumMechanics.CompositeBasis — Type

CompositeBasis(b::Basis...)

Create a basis for the composite system consisting of the systems corresponding to the given bases.

Examples

julia> CompositeBasis(GenericBasis(8), FockBasis(50, 20))
CompositeBasis{Tuple{GenericBasis{Int64}, FockBasis{Int64}}, Vector{Int64}}((GenericBasis{In
t64}(8), FockBasis{Int64}(31, 50, 20)), [8, 31])

QuantumMechanics.FockBasis — Type

FockBasis(cutoff::Integer, offset::Integer=0)

Create a basis for the Fock space starting at the offset and ending at the cutoff.

Examples

julia> FockBasis(10)
FockBasis{Int64}(11, 10, 0)

julia> FockBasis(10, 2)
FockBasis{Int64}(9, 10, 2)

QuantumMechanics.GenericBasis — Type

GenericBasis(d)

Create a generic basis with dimension d for a Hilbert space.

Examples

julia> GenericBasis(4)
GenericBasis{Int64}(4)

QuantumMechanics.SpinBasis — Type

SpinBasis(spin)

Create a basis for a spin system with total spin equal to spin.

The argument spin must be integer or half integer. It is then stored as a Rational.

Examples

julia> SpinBasis(1 // 2)
SpinBasis{1//2, Int64}(2, 1//2)

julia> SpinBasis(100)
SpinBasis{100//1, Int64}(201, 100//1)

Base.length — Method

length(b::Basis)

Return the total dimension of the Hilbert space spanned by the basis b.