QuantumMechanics

Documentation for QuantumMechanics.

QuantumMechanics.Bra

QuantumMechanics.CompositeBasis

QuantumMechanics.FockBasis

QuantumMechanics.GenericBasis

QuantumMechanics.Ket

QuantumMechanics.SpinBasis

Base.length

QuantumMechanics.dagger

QuantumMechanics.flatten

QuantumMechanics.partialtrace

QuantumMechanics.tensor

QuantumMechanics.Bra — Type

Bra(b::Basis, data::AbstractArray{Number})

Create a bra in the basis b with vector representation given by data.

The vector representation is converted to a vector whose elements are complex floats. It also need not be normalized.

Examples

julia> Bra(FockBasis(4), [1, 0, 0, 0, 0])
Bra{FockBasis{Int64}, ComplexF64}(FockBasis{Int64}((5,), 4, 0), ComplexF64[1.0 + 0.0im, 0.0 + 0.0im, 0.0 + 0.0im, 0.0 + 0.0im, 0.0 + 0.0im])

julia> Bra(SpinBasis(1 // 2), [1/sqrt(2), im/sqrt(2)])
Bra{SpinBasis{1//2, Int64}, ComplexF64}(SpinBasis{1//2, Int64}((2,), 1//2), ComplexF64[0.7071067811865475 + 0.0im, 0.0 + 0.7071067811865475im])

QuantumMechanics.CompositeBasis — Type

CompositeBasis(b::Basis...)

Create a basis for the composite system consisting of the systems corresponding to the given bases.

It is possible for a CompositeBasis to contain another CompositeBasis.

Examples

julia> composite_basis = CompositeBasis(GenericBasis(8), FockBasis(50, 20))
CompositeBasis{Tuple{GenericBasis{Int64}, FockBasis{Int64}}, Tuple{Int64, Int64}}((GenericBasis{Int64}((8,)), FockBasis{Int64}((31,), 50, 20)), (8, 31))

julia> CompositeBasis(composite_basis, SpinBasis(3//2))
CompositeBasis{Tuple{CompositeBasis{Tuple{GenericBasis{Int64}, FockBasis{Int64}}, Tuple{Int64, Int64}}, SpinBasis{3//2, Int64}}, Tuple{Int64, Int64}}((CompositeBasis{Tuple{GenericBasis{Int64}, FockBasis{Int64}}, Tuple{Int64, Int64}}((GenericBasis{Int64}((8,)), FockBasis{Int64}((31,), 50, 20)), (8, 31)), SpinBasis{3//2, Int64}((4,), 3//2)), (248, 4))

QuantumMechanics.FockBasis — Type

FockBasis(cutoff::Integer, offset::Integer=0)

Create a basis for the Fock space starting at the offset and ending at the cutoff.

Examples

julia> FockBasis(10)
FockBasis{Int64}((11,), 10, 0)

julia> FockBasis(10, 2)
FockBasis{Int64}((9,), 10, 2)

QuantumMechanics.GenericBasis — Type

GenericBasis(d)

Create a generic basis with dimension d for a Hilbert space.

Examples

julia> GenericBasis(4)
GenericBasis{Int64}((4,))

QuantumMechanics.Ket — Type

Ket(b::Basis, data::AbstractArray{Number})

Create a ket in the basis b with vector representation given by data.

The vector representation is converted to a vector whose elements are complex floats. It also need not be normalized.

Examples

julia> Ket(FockBasis(4), [1, 0, 0, 0, 0])
Ket{FockBasis{Int64}, ComplexF64}(FockBasis{Int64}((5,), 4, 0), ComplexF64[1.0 + 0.0im, 0.0 + 0.0im, 0.0 + 0.0im, 0.0 + 0.0im, 0.0 + 0.0im])

julia> Ket(SpinBasis(1 // 2), [1/sqrt(2), im/sqrt(2)])
Ket{SpinBasis{1//2, Int64}, ComplexF64}(SpinBasis{1//2, Int64}((2,), 1//2), ComplexF64[0.7071067811865475 + 0.0im, 0.0 + 0.7071067811865475im])

QuantumMechanics.SpinBasis — Type

SpinBasis(spin)

Create a basis for a spin system with total spin equal to spin.

The argument spin must be integer or half integer. It is then stored as a Rational.

Examples

julia> SpinBasis(1 // 2)
SpinBasis{1//2, Int64}((2,), 1//2)

julia> SpinBasis(100)
SpinBasis{100//1, Int64}((201,), 100//1)

Base.length — Method

length(b::Basis)

Return the total dimension of the Hilbert space spanned by the basis b.

QuantumMechanics.dagger — Method

dagger(ψ::StateVector)

Hermitian transpose of the state vector. Using dagger on a Ket or a Bra will respectively return a Bra or a Ket in the same basis as the input state, but whose data has been complex conjugated.

QuantumMechanics.flatten — Method

flatten(b::CompositeBasis)

Flatten nested composite basis. Works recursively.

Examples

julia> flatten(CompositeBasis(CompositeBasis(FockBasis(1), FockBasis(2)), FockBasis(3)))
CompositeBasis{Tuple{FockBasis{Int64}, FockBasis{Int64}, FockBasis{Int64}}, Tuple{Int64, Int64, Int64}}((FockBasis{Int64}((2,), 1, 0), FockBasis{Int64}((3,), 2, 0), FockBasis{Int64}((4,), 3, 0)), (2, 3, 4))

QuantumMechanics.partialtrace — Method

partialtrace(b::CompositeBasis, indices...)

Remove the bases with indices corresponding to the given indices from the composite bases b.

If the partial trace leaves multiple remaining bases, these will be return as a CompositeBasis. If only one basis remains after performing the partial trace, it will be returned as itself, i. e. not in a CompositeBasis.

Examples

julia> partialtrace(CompositeBasis(SpinBasis(1 // 2), FockBasis(10)), 2)
SpinBasis{1//2, Int64}((2,), 1//2)

julia> partialtrace(CompositeBasis(SpinBasis(1 // 2), SpinBasis(3 // 2), FockBasis(10)), 3)
CompositeBasis{Tuple{SpinBasis{1//2, Int64}, SpinBasis{3//2, Int64}}, Tuple{Int64, Int64}}((SpinBasis{1//2, Int64}((2,), 1//2), SpinBasis{3//2, Int64}((4,), 3//2)), (2, 4))

QuantumMechanics.tensor — Method

tensor(b::Basis...)

Create a basis for the composite system consisting of the systems corresponding to the given bases.

When using tensor, rather than CompositeBasis, any CompositeBasis will be expanded into its components, meaning that the CompositeBasis returned from tensor will never contain another CompositeBasis.

It is also possible to use to otimes symbol ⊗ as an infix operator.

Rasing a basis b to the power of N, where N is a positive integer results in the tensor product of N copies of b.

Examples

julia> composite_basis = tensor(GenericBasis(8), FockBasis(50, 20))
CompositeBasis{Tuple{GenericBasis{Int64}, FockBasis{Int64}}, Tuple{Int64, Int64}}((GenericBasis{Int64}((8,)), FockBasis{Int64}((31,), 50, 20)), (8, 31))

julia> tensor(composite_basis, SpinBasis(3//2))
CompositeBasis{Tuple{GenericBasis{Int64}, FockBasis{Int64}, SpinBasis{3//2, Int64}}, Tuple{Int64, Int64, Int64}}((GenericBasis{Int64}((8,)), FockBasis{Int64}((31,), 50, 20), SpinBasis{3//2, Int64}((4,), 3//2)), (8, 31, 4))

julia> GenericBasis(12) ⊗ FockBasis(25) ⊗ SpinBasis(1)
CompositeBasis{Tuple{GenericBasis{Int64}, FockBasis{Int64}, SpinBasis{1//1, Int64}}, Tuple{Int64, Int64, Int64}}((GenericBasis{Int64}((12,)), FockBasis{Int64}((26,), 25, 0), SpinBasis{1//1, Int64}((3,), 1//1)), (12, 26, 3))

julia> GenericBasis(5)^4
CompositeBasis{NTuple{4, GenericBasis{Int64}}, NTuple{4, Int64}}((GenericBasis{Int64}((5,)), GenericBasis{Int64}((5,)), GenericBasis{Int64}((5,)), GenericBasis{Int64}((5,))), (5, 5, 5, 5))